\name{windanomaly}
\docType{data}
\alias{windanomaly}
\title{
Eq. Pacific meridional wind anomaly index, Jan 1900 - June 2005
}
\description{
 This dataset gives the monthly ENSO meridional wind anomaly index for the region 12-2N, 160E-80W from January 1900 until June 2005.
}
\usage{
data("windanomaly")
}
\format{
	A vector of wind anomalies with length 1266.
}
\source{http://jisao.washington.edu/data_sets/eqpacmeridwindts/}

\keyword{datasets}
